feat: 에러코드 상수 분리
