+++ src/views/ErrorPage/constants/errorCode.tsx
@@ -1,0 +1,18 @@
+import { Ic404Back, Ic404Front, Ic404Ghost, Ic500Back, Ic500Cone, Ic500Front } from '../assets';
+
+const CODE_IMG = {
+  CODE404:
+    <>
+      <Ic404Front/>
+      <Ic404Ghost/>
+      <Ic404Back/>
+    </>,
+  CODE500: 
+    <>
+      <Ic500Front />
+      <Ic500Cone />
+      <Ic500Back />
+    </>,
+};
+
+export default CODE_IMG;
